src/test/kotlin/year_2024/day_25/Task1Test.kt
package year_2024.day_25

import org.junit.jupiter.api.Test
import java.io.File
import kotlin.test.assertEquals

class Task1Test {

    @Test
    fun `should successfully pass task example`() {
        val input = """
            #####
            .####
            .####
            .####
            .#.#.
            .#...
            .....
            
            #####
            ##.##
            .#.##
            ...##
            ...#.
            ...#.
            .....
            
            .....
            #....
            #....
            #...#
            #.#.#
            #.###
            #####
            
            .....
            .....
            #.#..
            ###..
            ###.#
            ###.#
            #####
            
            .....
            .....
            .....
            #....
            #.#..
            #.#.#
            #####
        """.trimIndent()
        val (locks, keys) = Task1.transform(input)
        assertEquals(3, Task1.solve(locks, keys))
    }

    @Test
    fun `should successfully solve the real task`() {
        val input = File("src/test/resources/year_2024/day_25/TaskData.txt").readText()
        val (locks, keys) = Task1.transform(input)
        assertEquals(3077, Task1.solve(locks, keys))
    }
}